ChemConverter/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.0
pydantic==2.8.2
python-dotenv==1.0.1
cirpy==1.0.2
requests>=2.32.0
pydantic-settings>=2.3.4
numpy==1.26.4
rdkit-pypi==2022.9.5
